ADD: Eliminado y actualizacion de usuario. Encriptacion de contrasenias.

+++ PNCController/src/main/java/com/renegz/pnccontroller/services/implementations/BookServiceImpl.java
@@ -5,13 +5,12 @@ import com.renegz.pnccontroller.domain.entities.Book;
 import com.renegz.pnccontroller.domain.entities.Category;
 import com.renegz.pnccontroller.repositories.BookRepository;
 import com.renegz.pnccontroller.services.BookService;
-import jakarta.transaction.TransactionScoped;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service //Ahora ya lo puede ocupas el entityManager para poder inyectarlo
+@Service //Ahora ya lo puede ocupar el entityManager para poder inyectarlo
 public class BookServiceImpl implements BookService {
 
     private final BookRepository bookRepository;
@@ -23,13 +22,13 @@ public class BookServiceImpl implements BookService {
     @Override
     @Transactional(rollbackOn = Exception.class)
     public void save(SaveBookDTO info, Category category) {
-        Book book = this.findByIsbn(info.getISBN());
+        Book book = this.findByIsbn(info.getIsbn());
 
         if (book == null) {
             book = new Book();
         }
 
-        book.setIsbn(info.getISBN());
+        book.setIsbn(info.getIsbn());
         book.setTitle(info.getTitle());
         book.setCategory(category);
         bookRepository.save(book);
@@ -46,7 +45,12 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
+    @Transactional(rollbackOn = Exception.class)
     public void deleteByIsbn(String isbn) {
-        bookRepository.deleteByIsbn(isbn);
+        Book book = this.findByIsbn(isbn);
+
+        if (book != null) {
+            bookRepository.deleteByIsbn(isbn);
+        }
     }
 }
